directional edge fade. cylinderAB implementation

+++ tracevis/headers/plotted_graph.h
@@ -173,7 +173,7 @@ protected:
 	int maxA = 0, maxB = 0;
 
 	//keep track of which a,b coords are occupied
-	std::map<pair<int, int>, bool> usedCoords;
+	std::map<pair<float, float>, bool> usedCoords;
 	vector<pair<MEM_ADDRESS, NODEINDEX>> *callStack;
 	vector<pair<MEM_ADDRESS, NODEINDEX>> mainCallStack;
 	vector<pair<MEM_ADDRESS, NODEINDEX>> previewCallStack;
